Moved PictureBoard to pages + file adjustments
